requirements.txt
umap-learn
numpy
matplotlib